refactor(design-system): Move resize types to a separate file
This will prevent long import statements in later commits.
