+++ packages/design-system/src/types/index.ts
@@ -9,3 +9,4 @@ export * from './select';
 export * from './user';
 export * from './keyboardshortcut';
 export * from './node-creator-node';
+export * from './resize';
